+++ Network/API/Telegram/Bot/Object/Update/Message/Content/File/Document.hs
@@ -12,8 +12,7 @@ import "text" Data.Text (Text)
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.URI (URI)
 import Network.API.Telegram.Bot.Property (Identifiable (Identificator, ident))
 
-data Document = Document URI (Maybe Text) (Maybe Text) (Maybe Int)
-	deriving Show
+data Document = Document URI (Maybe Text) (Maybe Text) (Maybe Int) deriving Show
 
 instance Identifiable Document where
 	type Identificator Document = URI